:zap: navigation bar를 hook으로 변경 #4
